coba2/requirements.txt
# =========================
# File: requirements.txt
# =========================
streamlit
opencv-python-headless
numpy
scipy
pillow
pandas